requirements-dev.txt
# Development dependencies with specific versions for reproducibility
-r requirements.txt  # Include production dependencies

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mypy==0.10.3
pytest-env==1.1.1

# Code quality
black==23.7.0
isort==5.12.0
mypy==1.4.1
types-PyYAML==6.0.12.11
types-python-dateutil==2.8.19.14

# Pydantic and related
pydantic>=1.10.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
PyYAML>=6.0.0
